requirements.txt
gradio==3.50.0
llama-cpp-python==0.2.23
diffusers==0.25.0
torch==2.0.1+cpu